requirements.txt
numpy==1.18.1
pandas==1.0.1
asn1crypto==0.24.0
bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
dj-database-url==0.4.2
Django==1.11.15
django-registration==2.2
gunicorn==19.7.1
idna==2.8
keen==0.5.1
lxml==4.2.5
MarkupSafe==1.1.0
pymongo==3.10.1
paramiko==2.4.2
psycopg2==2.7.1
pyasn1==0.4.5
pycparser==2.19
pycryptodome==3.6.6
PyNaCl==1.1.2
PySocks==1.6.8
pytz==2017.2
requests==2.21.0
six==1.10.0
urllib3==1.24.1
whitenoise==3.3.0
